# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_locde0_sportudei-ukma_backend_internal_dto.AdminGalleryAlbumResponse:
    properties:
      cover_image_path:
        type: string
      id:
        type: integer
      is_published:
        type: boolean
      title:
        type: string
    type: object
  github_com_locde0_sportudei-ukma_backend_internal_dto.AdminGalleryAlbumsListResponse:
    properties:
      albums:
        items:
          $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.AdminGalleryAlbumResponse'
        type: array
    type: object
...
  github_com_locde0_sportudei-ukma_backend_internal_dto.CreateContactRequest:
    properties:
      displayOrder:
        minimum: 0
        type: integer
      name:
        maxLength: 50
        type: string
      platform:
        enum:
        - telegram
        - instagram
        - facebook
        - email
        type: string
      url:
        type: string
    required:
    - name
    - platform
    - url
    type: object
...
  github_com_locde0_sportudei-ukma_backend_internal_dto.GalleryPhotoResponse:
    properties:
      display_order:
        type: integer
      id:
        type: integer
      image_path:
        type: string
    type: object
  github_com_locde0_sportudei-ukma_backend_internal_dto.GalleryPhotosListResponse:
    properties:
      photos:
        items:
          $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.GalleryPhotoResponse'
        type: array
    type: object
...
  github_com_locde0_sportudei-ukma_backend_internal_dto.PublicGalleryAlbumResponse:
    properties:
      cover_image_path:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  github_com_locde0_sportudei-ukma_backend_internal_dto.PublicGalleryAlbumsListResponse:
    properties:
      albums:
        items:
          $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.PublicGalleryAlbumResponse'
        type: array
    type: object
...
  github_com_locde0_sportudei-ukma_backend_internal_dto.UpdateContactOrderRequest:
    properties:
      displayOrder:
        minimum: 0
        type: integer
    type: object
  github_com_locde0_sportudei-ukma_backend_internal_dto.UpdateContactRequest:
    properties:
      name:
        maxLength: 50
        type: string
      platform:
        enum:
        - telegram
        - instagram
        - facebook
        - email
        type: string
      url:
        type: string
    required:
    - name
    - platform
    - url
    type: object
...
  github_com_locde0_sportudei-ukma_backend_internal_dto.UpdateGalleryAlbumRequest:
    properties:
      cover_image_path:
        type: string
      is_published:
        type: boolean
      photos:
        items:
          $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.UpdateGalleryPhotoRequest'
        type: array
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - title
    type: object
  github_com_locde0_sportudei-ukma_backend_internal_dto.UpdateGalleryPhotoRequest:
    properties:
      display_order:
        minimum: -1
        type: integer
      id:
        type: integer
    required:
    - id
    type: object
...
info:
  contact: {}
  description: sportudei api server
  title: sportudei api
  version: "1.0"
paths:
  /api/admin/contacts:
    post:
      consumes:
      - application/json
      description: Create a new contact link
      parameters:
      - description: Contact data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.CreateContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - BearerAuth: []
      summary: Create contact
      tags:
      - admin-contacts
  /api/admin/contacts/{id}:
    delete:
      description: Delete an existing contact link by ID
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Delete contact
      tags:
      - admin-contacts
    put:
      consumes:
      - application/json
      description: Update an existing contact link
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.UpdateContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Update contact
      tags:
      - admin-contacts
  /api/admin/contacts/{id}/order:
    put:
      consumes:
      - application/json
      description: Update the display order of a contact
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      - description: New order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.UpdateContactOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Update contact order
      tags:
      - admin-contacts
...
  /api/admin/gallery:
    get:
      description: List all albums for admin view with pagination
      parameters:
      - default: 10
        description: Pagination limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.AdminGalleryAlbumsListResponse'
      security:
      - BearerAuth: []
      summary: List admin albums
      tags:
      - admin-gallery
    post:
      consumes:
      - multipart/form-data
      description: Create a new gallery album with a cover photo
      parameters:
      - description: CreateGalleryAlbumRequest JSON string
        in: formData
        name: payload
        required: true
        type: string
      - description: Cover photo
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - BearerAuth: []
      summary: Create album
      tags:
      - admin-gallery
  /api/admin/gallery/{id}:
    delete:
      description: Delete an existing gallery album by ID
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Delete album
      tags:
      - admin-gallery
    get:
      description: Get full album details for admin by ID
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.AdminGalleryAlbumResponse'
      security:
      - BearerAuth: []
      summary: Get admin album
      tags:
      - admin-gallery
    put:
      consumes:
      - application/json
      description: Update an existing gallery album and its photos
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.UpdateGalleryAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Update album
      tags:
      - admin-gallery
  /api/admin/gallery/{id}/photos:
    post:
      consumes:
      - multipart/form-data
      description: Upload an additional photo for an album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Photo to upload
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Upload album photo
      tags:
      - admin-gallery
...
  /api/gallery:
    get:
      description: List published albums for public view with pagination
      parameters:
      - default: 6
        description: Pagination limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.PublicGalleryAlbumsListResponse'
      summary: List public albums
      tags:
      - public-gallery
  /api/gallery/{id}:
    get:
      description: Get album details for public view by ID
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.PublicGalleryAlbumResponse'
      summary: Get public album
      tags:
      - public-gallery
  /api/gallery/{id}/photos:
    get:
      description: Get photos of an album with pagination
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - default: 10
        description: Pagination limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.GalleryPhotosListResponse'
      summary: Get album photos
      tags:
      - public-gallery
...
swagger: "2.0"